Move relocation endpoints, serializer and models
