+++ src/sentry/relocation/api/endpoints/abort.py
@@ -9,7 +9,7 @@ from sentry.api.base import Endpoint, region_silo_endpoint
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.permissions import SuperuserOrStaffFeatureFlaggedPermission
 from sentry.api.serializers import serialize
-from sentry.models.relocation import Relocation
+from sentry.relocation.models.relocation import Relocation
 
 ERR_NOT_ABORTABLE_STATUS = (
     "Relocations can only be cancelled if they are not yet complete; this relocation is `SUCCESS`."
