+++ src/sentry/relocation/api/endpoints/cancel.py
@@ -9,11 +9,11 @@ from rest_framework.response import Response
 from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import Endpoint, region_silo_endpoint
-from sentry.api.endpoints.relocations import ERR_UNKNOWN_RELOCATION_STEP
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.permissions import SuperuserOrStaffFeatureFlaggedPermission
 from sentry.api.serializers import serialize
-from sentry.models.relocation import Relocation
+from sentry.relocation.api.endpoints import ERR_UNKNOWN_RELOCATION_STEP
+from sentry.relocation.models.relocation import Relocation
 
 ERR_NOT_CANCELLABLE_STATUS = (
     "Relocations can only be cancelled if they are not yet complete; this relocation is `SUCCESS`."
