+++ tests/sentry/db/models/test_base.py
@@ -11,12 +11,6 @@ from sentry.models.importchunk import (
     RegionImportChunk,
 )
 from sentry.models.projecttemplate import ProjectTemplate
-from sentry.models.relocation import (
-    Relocation,
-    RelocationFile,
-    RelocationValidation,
-    RelocationValidationAttempt,
-)
 from sentry.models.transaction_threshold import (
     ProjectTransactionThreshold,
     ProjectTransactionThresholdOverride,
@@ -24,6 +18,12 @@ from sentry.models.transaction_threshold import (
 from sentry.notifications.models.notificationsettingbase import NotificationSettingBase
 from sentry.notifications.models.notificationsettingoption import NotificationSettingOption
 from sentry.notifications.models.notificationsettingprovider import NotificationSettingProvider
+from sentry.relocation.models.relocation import (
+    Relocation,
+    RelocationFile,
+    RelocationValidation,
+    RelocationValidationAttempt,
+)
 from sentry.sentry_apps.models import SentryAppInstallationForProvider
 from sentry.uptime.models import ProjectUptimeSubscription, UptimeSubscription
