+++ tests/sentry/relocation/api/endpoints/test_pause.py
@@ -1,13 +1,15 @@
 from datetime import datetime, timezone
 from uuid import uuid4
 
-from sentry.api.endpoints.relocations import ERR_UNKNOWN_RELOCATION_STEP
-from sentry.api.endpoints.relocations.cancel import (
-    ERR_COULD_NOT_CANCEL_RELOCATION,
-    ERR_COULD_NOT_CANCEL_RELOCATION_AT_STEP,
-    ERR_NOT_CANCELLABLE_STATUS,
+from sentry.relocation.api.endpoints import (
+    ERR_COULD_NOT_PAUSE_RELOCATION_AT_STEP,
+    ERR_UNKNOWN_RELOCATION_STEP,
 )
-from sentry.models.relocation import Relocation
+from sentry.relocation.api.endpoints.pause import (
+    ERR_COULD_NOT_PAUSE_RELOCATION,
+    ERR_NOT_PAUSABLE_STATUS,
+)
+from sentry.relocation.models.relocation import Relocation
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.helpers.options import override_options
 from sentry.utils.relocation import OrderedTask
@@ -15,8 +17,8 @@ from sentry.utils.relocation import OrderedTask
 TEST_DATE_ADDED = datetime(2023, 1, 23, 1, 23, 45, tzinfo=timezone.utc)
 
 
-class CancelRelocationTest(APITestCase):
-    endpoint = "sentry-api-0-relocations-cancel"
+class PauseRelocationTest(APITestCase):
+    endpoint = "sentry-api-0-relocations-pause"
     method = "put"
 
     def setUp(self):
@@ -25,6 +27,7 @@ class CancelRelocationTest(APITestCase):
             email="owner", is_superuser=False, is_staff=True, is_active=True
         )
         self.superuser = self.create_user(is_superuser=True)
+        self.staff_user = self.create_user(is_staff=True)
         self.relocation: Relocation = Relocation.objects.create(
             date_added=TEST_DATE_ADDED,
             creator_id=self.superuser.id,
@@ -40,56 +43,33 @@ class CancelRelocationTest(APITestCase):
         )
 
     @override_options({"staff.ga-rollout": True})
-    def test_good_staff_cancel_in_progress_at_next_step(self):
-        staff_user = self.create_user(is_staff=True)
-        self.login_as(user=staff_user, staff=True)
+    def test_good_staff_pause_asap(self):
+        self.login_as(user=self.staff_user, staff=True)
         response = self.get_success_response(self.relocation.uuid, status_code=200)
 
         assert response.data["status"] == Relocation.Status.IN_PROGRESS.name
         assert response.data["step"] == Relocation.Step.PREPROCESSING.name
-        assert response.data["scheduledCancelAtStep"] == Relocation.Step.VALIDATING.name
+        assert response.data["scheduledPauseAtStep"] == Relocation.Step.VALIDATING.name
 
-    def test_good_cancel_in_progress_at_next_step(self):
+    def test_good_pause_asap(self):
         self.login_as(user=self.superuser, superuser=True)
         response = self.get_success_response(self.relocation.uuid, status_code=200)
 
         assert response.data["status"] == Relocation.Status.IN_PROGRESS.name
         assert response.data["step"] == Relocation.Step.PREPROCESSING.name
-        assert response.data["scheduledCancelAtStep"] == Relocation.Step.VALIDATING.name
+        assert response.data["scheduledPauseAtStep"] == Relocation.Step.VALIDATING.name
 
-    def test_good_cancel_paused_at_next_step(self):
-        self.login_as(user=self.superuser, superuser=True)
-        self.relocation.status = Relocation.Status.PAUSE.value
-        self.relocation.save()
-        response = self.get_success_response(self.relocation.uuid, status_code=200)
-
-        assert response.data["status"] == Relocation.Status.PAUSE.name
-        assert response.data["step"] == Relocation.Step.PREPROCESSING.name
-        assert response.data["scheduledCancelAtStep"] == Relocation.Step.VALIDATING.name
-
-    def test_good_cancel_in_progress_at_specified_step(self):
+    def test_good_pause_at_next_step(self):
         self.login_as(user=self.superuser, superuser=True)
         response = self.get_success_response(
-            self.relocation.uuid, atStep=Relocation.Step.IMPORTING.name, status_code=200
+            self.relocation.uuid, atStep=Relocation.Step.VALIDATING.name, status_code=200
         )
 
         assert response.data["status"] == Relocation.Status.IN_PROGRESS.name
         assert response.data["step"] == Relocation.Step.PREPROCESSING.name
-        assert response.data["scheduledCancelAtStep"] == Relocation.Step.IMPORTING.name
+        assert response.data["scheduledPauseAtStep"] == Relocation.Step.VALIDATING.name
 
-    def test_good_cancel_paused_at_specified_step(self):
-        self.login_as(user=self.superuser, superuser=True)
-        self.relocation.status = Relocation.Status.PAUSE.value
-        self.relocation.save()
-        response = self.get_success_response(
-            self.relocation.uuid, atStep=Relocation.Step.IMPORTING.name, status_code=200
-        )
-
-        assert response.data["status"] == Relocation.Status.PAUSE.name
-        assert response.data["step"] == Relocation.Step.PREPROCESSING.name
-        assert response.data["scheduledCancelAtStep"] == Relocation.Step.IMPORTING.name
-
-    def test_good_cancel_at_future_step(self):
+    def test_good_pause_at_future_step(self):
         self.login_as(user=self.superuser, superuser=True)
         response = self.get_success_response(
             self.relocation.uuid, atStep=Relocation.Step.NOTIFYING.name, status_code=200
@@ -97,45 +77,37 @@ class CancelRelocationTest(APITestCase):
 
         assert response.data["status"] == Relocation.Status.IN_PROGRESS.name
         assert response.data["step"] == Relocation.Step.PREPROCESSING.name
-        assert response.data["scheduledCancelAtStep"] == Relocation.Step.NOTIFYING.name
+        assert response.data["scheduledPauseAtStep"] == Relocation.Step.NOTIFYING.name
 
-    def test_good_already_cancelled(self):
+    def test_good_already_paused(self):
         self.login_as(user=self.superuser, superuser=True)
-        self.relocation.scheduled_cancel_at_step = Relocation.Step.IMPORTING.value
+        self.relocation.status = Relocation.Status.PAUSE.value
         self.relocation.save()
         response = self.get_success_response(
             self.relocation.uuid, atStep=Relocation.Step.IMPORTING.name, status_code=200
         )
 
-        assert response.data["status"] == Relocation.Status.IN_PROGRESS.name
+        assert response.data["status"] == Relocation.Status.PAUSE.name
         assert response.data["step"] == Relocation.Step.PREPROCESSING.name
-        assert response.data["scheduledCancelAtStep"] == Relocation.Step.IMPORTING.name
-
-    def test_good_already_failed(self):
-        self.login_as(user=self.superuser, superuser=True)
-        self.relocation.status = Relocation.Status.FAILURE.value
-        self.relocation.save()
-        response = self.get_success_response(
-            self.relocation.uuid, atStep=Relocation.Step.PREPROCESSING.name, status_code=200
-        )
-
-        assert response.data["status"] == Relocation.Status.FAILURE.name
-        assert response.data["step"] == Relocation.Step.PREPROCESSING.name
-        assert not response.data["scheduledCancelAtStep"]
+        assert response.data["scheduledPauseAtStep"] == Relocation.Step.IMPORTING.name
 
     def test_bad_not_found(self):
         self.login_as(user=self.superuser, superuser=True)
         does_not_exist_uuid = uuid4().hex
-        self.get_error_response(does_not_exist_uuid, status_code=404)
+        response = self.client.put(f"/api/0/relocations/{str(does_not_exist_uuid)}/pause/")
 
-    def test_bad_already_succeeded(self):
+        assert response.status_code == 404
+
+    def test_bad_already_completed(self):
         self.login_as(user=self.superuser, superuser=True)
-        self.relocation.status = Relocation.Status.SUCCESS.value
+        self.relocation.status = Relocation.Status.FAILURE.value
         self.relocation.save()
         response = self.get_error_response(self.relocation.uuid, status_code=400)
 
         assert response.data.get("detail") is not None
-        assert response.data.get("detail") == ERR_NOT_CANCELLABLE_STATUS
+        assert response.data.get("detail") == ERR_NOT_PAUSABLE_STATUS.substitute(
+            status=Relocation.Status.FAILURE.name
+        )
 
     def test_bad_invalid_step(self):
         self.login_as(user=self.superuser, superuser=True)
@@ -155,7 +127,7 @@ class CancelRelocationTest(APITestCase):
         )
 
         assert response.data.get("detail") is not None
-        assert response.data.get("detail") == ERR_COULD_NOT_CANCEL_RELOCATION_AT_STEP.substitute(
+        assert response.data.get("detail") == ERR_COULD_NOT_PAUSE_RELOCATION_AT_STEP.substitute(
             step=Relocation.Step.UNKNOWN.name
         )
 
@@ -166,7 +138,9 @@ class CancelRelocationTest(APITestCase):
         )
 
         assert response.data.get("detail") is not None
-        assert response.data.get("detail") == ERR_COULD_NOT_CANCEL_RELOCATION
+        assert response.data.get("detail") == ERR_COULD_NOT_PAUSE_RELOCATION_AT_STEP.substitute(
+            step=Relocation.Step.PREPROCESSING.name
+        )
 
     def test_bad_past_step(self):
         self.login_as(user=self.superuser, superuser=True)
@@ -175,7 +149,7 @@ class CancelRelocationTest(APITestCase):
         )
 
         assert response.data.get("detail") is not None
-        assert response.data.get("detail") == ERR_COULD_NOT_CANCEL_RELOCATION_AT_STEP.substitute(
+        assert response.data.get("detail") == ERR_COULD_NOT_PAUSE_RELOCATION_AT_STEP.substitute(
             step=Relocation.Step.UPLOADING.name
         )
 
@@ -186,7 +160,7 @@ class CancelRelocationTest(APITestCase):
         )
 
         assert response.data.get("detail") is not None
-        assert response.data.get("detail") == ERR_COULD_NOT_CANCEL_RELOCATION_AT_STEP.substitute(
+        assert response.data.get("detail") == ERR_COULD_NOT_PAUSE_RELOCATION_AT_STEP.substitute(
             step=Relocation.Step.COMPLETED.name
         )
 
@@ -197,7 +171,7 @@ class CancelRelocationTest(APITestCase):
         response = self.get_error_response(self.relocation.uuid, status_code=400)
 
         assert response.data.get("detail") is not None
-        assert response.data.get("detail") == ERR_COULD_NOT_CANCEL_RELOCATION
+        assert response.data.get("detail") == ERR_COULD_NOT_PAUSE_RELOCATION
 
     def test_bad_no_auth(self):
         self.get_error_response(self.relocation.uuid, status_code=401)
